Add members bottom nav layout variant
